requirements.txt
fyers-apiv3
pandas
pandas-ta
pyotp
requests
flask
gunicorn
setuptools
python-dotenv
